fix(migrations): repair five SQL bugs surfaced by db push to remote
Eleven migrations were sitting locally in the migrations folder but
had never been applied to remote. Pushing them in order surfaced bugs
that local development wasn't exercising. Fixed in place so the
history is clean and `supabase db push` is a no-op going forward.

1. 20260411120000_mma_boxing_upgrade → 20260411120001_mma_boxing_upgrade.sql
   • Renamed: another file already used 20260411120000, causing a
     primary-key collision in supabase_migrations.
   • Fixed boxing_prediction_factors.prediction_id type: text → bigint
     (boxing_predictions.prediction_id is bigint on remote; the FK
     declared text and rejected with SQLSTATE 42804).
   • Replaced two `CREATE POLICY IF NOT EXISTS` blocks (Postgres has
     no such syntax) with idempotent DROP-then-CREATE pairs so the
     migration is re-runnable.

2. 20260430100000_accuracy_feedback_loop.sql
   • prediction_history.resolved_at → settled_at (column never
     existed; the table has settled_at instead).

3. 20260430300000_ml_performance_rollups.sql
   • Same resolved_at → settled_at substitution (3 places).
   • Removed dead `LEFT JOIN LATERAL (select ph.sport)` from
     analytics_model_performance_by_sport — leftover from a refactor
     that made the GROUP BY ambiguous.

4. 20260502000000_ml_training_samples.sql
   • Replaced `current_date` with sentinel `'1970-01-01'::date` in a
     UNIQUE INDEX expression — Postgres requires IMMUTABLE functions
     in index keys, and current_date is STABLE.

5. 20260511000000_analytics_clv.sql
   • Cast `percentile_cont(...)` result to numeric before round() —
     percentile_cont returns double precision and there's no
     round(double precision, integer) overload.

+++ supabase/migrations/20260511000000_analytics_clv.sql
@@ -86,7 +86,9 @@ as $$
     )                                                                            as pct_beating_close,
     round((select avg(clv_pp) from apples), 2)                                   as avg_clv_pp,
     round(
-      (select percentile_cont(0.5) within group (order by clv_pp) from apples),
+      -- percentile_cont returns double precision regardless of input
+      -- type; round(double precision, int) doesn't exist, so cast.
+      (select percentile_cont(0.5) within group (order by clv_pp) from apples)::numeric,
       2
     )                                                                            as median_clv_pp,
     (select count(*) from apples_7d)                                             as total_7d,
